+++ Zoop/Core/ZoopPreviewCoordinator.cs
@@ -16,8 +16,11 @@ namespace ZoopMod.Zoop.Core;
 /// </summary>
 internal sealed class ZoopPreviewCoordinator(ZoopPreviewValidator previewValidator)
 {
-  private readonly ZoopSmallGridCoordinator smallGridCoordinator = new(previewValidator);
-  private readonly ZoopBigGridCoordinator bigGridCoordinator = new(previewValidator);
+  private readonly IZoopGridStrategy[] gridStrategies =
+  [
+    new ZoopSmallGridCoordinator(previewValidator),
+    new ZoopBigGridCoordinator(previewValidator),
+  ];
 
   private ZoopDraft activeDraft;
   private ZoopPreviewCache activePreviewCache;
@@ -103,19 +106,18 @@ internal sealed class ZoopPreviewCoordinator(ZoopPreviewValidator previewValidat
     }
   }
 
-  internal static Vector3? GetCurrentMouseGridPosition()
+  internal IZoopGridStrategy FindStrategy(Structure cursor)
+    => cursor != null ? Array.Find(gridStrategies, s => s.Matches(cursor)) : null;
+
+  internal Vector3? GetCurrentMouseGridPosition()
   {
-    if (InventoryManager.ConstructionCursor == null)
+    var cursor = InventoryManager.ConstructionCursor;
+    if (cursor == null)
     {
       return null;
     }
 
-    if (InventoryManager.ConstructionCursor is Wall)
-    {
-      return InventoryManager.ConstructionCursor.ThingTransformPosition;
-    }
-
-    return InventoryManager.ConstructionCursor.GetLocalGrid().ToVector3();
+    return FindStrategy(cursor)?.GetCursorPosition(cursor);
   }
 
   private IEnumerator PreviewLoop(ZoopDraft draft, ZoopPreviewCache previewCache, InventoryManager inventoryManager)
@@ -159,13 +161,11 @@ internal sealed class ZoopPreviewCoordinator(ZoopPreviewValidator previewValidat
 
     draft.HasError = false;
 
-    if (IsZoopingSmallGrid())
+    var cursor = InventoryManager.ConstructionCursor;
+    var strategy = Array.Find(gridStrategies, s => s.Matches(cursor));
+    if (strategy != null)
     {
-      await smallGridCoordinator.UpdatePreview(draft, previewCache, inventoryManager, currentPos);
-    }
-    else if (IsZoopingBigGrid())
-    {
-      await bigGridCoordinator.UpdatePreview(draft, previewCache, inventoryManager, currentPos);
+      await strategy.UpdatePreview(draft, previewCache, inventoryManager, currentPos);
     }
 
     UpdateFullFidelityPieces(draft);
@@ -182,16 +182,6 @@ internal sealed class ZoopPreviewCoordinator(ZoopPreviewValidator previewValidat
     return previewDirty ||
            !lastPreviewCursorPosition.HasValue ||
            !ZoopPositionUtility.IsSameZoopPosition(lastPreviewCursorPosition.Value, currentPos);
-  }
-
-  private static bool IsZoopingSmallGrid()
-  {
-    return InventoryManager.ConstructionCursor is SmallGrid;
-  }
-
-  private static bool IsZoopingBigGrid()
-  {
-    return InventoryManager.ConstructionCursor is LargeStructure;
   }
 
   private void UpdateFullFidelityPieces(ZoopDraft draft)
