+++ Zoop/Placement/ZoopBigGridCoordinator.cs
@@ -1,9 +1,11 @@
+using System;
 using Assets.Scripts.Inventory;
 using Assets.Scripts.Objects;
 using Cysharp.Threading.Tasks;
 using UnityEngine;
 using ZoopMod.Zoop.Core;
 using ZoopMod.Zoop.Planning;
+using ZoopMod.Zoop.Planning.BigGrid;
 using ZoopMod.Zoop.Preview;
 
 namespace ZoopMod.Zoop.Placement;
@@ -11,8 +13,12 @@ namespace ZoopMod.Zoop.Placement;
 /// <summary>
 /// Owns the large-grid zoop preview flow from wall-plane planning through preview placement.
 /// </summary>
-internal sealed class ZoopBigGridCoordinator(ZoopPreviewValidator previewValidator)
+internal sealed class ZoopBigGridCoordinator(ZoopPreviewValidator previewValidator) : IZoopGridStrategy
 {
+  public bool Matches(Structure cursor) => cursor is LargeStructure;
+  public bool SupportsWaypoints => false;
+  public Vector3? GetCursorPosition(Structure cursor) => cursor.ThingTransformPosition;
+
   private sealed class BigGridPreviewLayoutAdapter(ZoopPreviewValidator previewValidator)
     : IBigGridPreviewLayoutAdapter
   {
@@ -40,7 +46,7 @@ internal sealed class ZoopBigGridCoordinator(ZoopPreviewValidator previewValidat
     var startPos = draft.Waypoints[0];
     var endPos = ClampWallZoopPositionToStartPlane(draft, startPos, currentPos);
 
-    var plane = ZoopPathPlanner.BuildBigGridPlane(startPos, endPos);
+    var plane = BuildBigGridPlane(startPos, endPos);
 
     await UniTask.SwitchToMainThread(); // Switch to main thread for Unity API calls
 
@@ -104,4 +110,42 @@ internal sealed class ZoopBigGridCoordinator(ZoopPreviewValidator previewValidat
 
     return targetPos;
   }
+
+  private static ZoopPlane BuildBigGridPlane(Vector3 startPos, Vector3 endPos)
+  {
+    var plane = new ZoopPlane();
+    CalculateZoopPlane(startPos, endPos, plane);
+
+    if (!ZoopPositionUtility.IsSameZoopPosition(startPos, endPos))
+    {
+      return plane;
+    }
+
+    plane.Count = (direction1: 1 + (int)(Math.Abs(startPos.x - endPos.x) / 2),
+      direction2: 1 + (int)(Math.Abs(startPos.x - endPos.x) / 2));
+    plane.Increasing = (direction1: startPos.x < endPos.x, direction2: startPos.y < endPos.y);
+    plane.Directions = (direction1: ZoopDirection.x, direction2: ZoopDirection.y);
+    return plane;
+  }
+
+  private static void CalculateZoopPlane(Vector3 startPos, Vector3 endPos, ZoopPlane plane)
+  {
+    var ax = (dir: ZoopDirection.x, abs: Math.Abs(endPos.x - startPos.x) / 2, s: startPos.x, e: endPos.x);
+    var ay = (dir: ZoopDirection.y, abs: Math.Abs(endPos.y - startPos.y) / 2, s: startPos.y, e: endPos.y);
+    var az = (dir: ZoopDirection.z, abs: Math.Abs(endPos.z - startPos.z) / 2, s: startPos.z, e: endPos.z);
+
+    if (ay.abs > ax.abs) (ax, ay) = (ay, ax);
+    if (az.abs > ax.abs) (ax, az) = (az, ax);
+    if (az.abs > ay.abs) (ay, _) = (az, ay);
+
+    var (cnt1, inc1) = AxisMetrics(ax.s, ax.e);
+    var (cnt2, inc2) = AxisMetrics(ay.s, ay.e);
+
+    plane.Directions = (direction1: ax.dir, direction2: ay.dir);
+    plane.Count = (direction1: cnt1, direction2: cnt2);
+    plane.Increasing = (direction1: inc1, direction2: inc2);
+  }
+
+  private static (int cnt, bool inc) AxisMetrics(float s, float e)
+    => (1 + (int)(Math.Abs(s - e) / 2), s < e);
 }
