+++ Zoop/Planning/BigGrid/ZoopPlane.cs
@@ -1,4 +1,6 @@
-namespace ZoopMod.Zoop.Planning;
+using ZoopMod.Zoop.Planning;
+
+namespace ZoopMod.Zoop.Planning.BigGrid;
 
 public class ZoopPlane
 {
