+++ ZoopMod.info
@@ -2,7 +2,7 @@
     "_name":  "ZoopMod",
     "_author":  "Nivvdiy, MerGatto",
     "_description":  "ZoopMod",
-    "_version":  "2026.03.04",
+    "_version":  "2026.04.04",
     "_unityVersion":  "2022.3.7f1",
     "_platforms":  1,
     "_content":  4,
